Add files via upload
There are two bulks operation for one table, the reason of that is that task guideline wat to prepare data in two snapshots. 
This solution allows user to check if SCD 2 is working correctly in the ETL process. 
Data generator for each of the table and for each snapshot is provided on the project respository